imported generate-poracle-v3-forms-form-rocketmad.py
generate-poracle-v3-forms-form-rocketmad.py reads RocketMAD’s
“static/data/pokemon.json” and generates Poracle’s V3 “src/util/forms.json” from
it.

There are several sanity checks and also a hard-coded translation of certain
form names.

If for a given Pokémon number, multiple form IDs map to the same form name, only
the one with the lowest ID is considered.
As of now, this seems to happen only for RocketMAD’s “normal” form name (which
is the empty) string, but which is also found on certain other forms, which are
presumably however just unnamed event forms (for which specifc tracking wouldn’t
make much sense anyway).

There’s also a command line option to ignore arguments.

Signed-off-by: Philippe Cerfon <philcerf@gmail.com>