updated German locale
This was made besed upon config/locale/fr.json, however, with all quest texts
removed.

As far as applicable, I've used the “offical“ spellings as found in the game.
This goes so far, that I've translated “genderless” with „geschlechtunbekannt“,
which no-one would use in German, however, „geschlechtunbekannt“ is used by the
game as search-keyword, so I stayed with that.
The official name of the “Adventure Hat 2020” form of Pikachu seems to be
“Explorer Pikachu” respectively “Forscher Pikachu” in German.

Also, as described in https://github.com/KartulUdus/PoracleJS/issues/320 I lack
context for a number of strings.
But in a second iteration, I think I've probably identified most of them
correctly and adapted any spellings as necessary.

Commands should IMO be written lower-case, even in German.
As far as I could identify strings belonging to commands, I left them
lower-case.

Further, I've tried to really translate most things, even if it’s strange.
For example, most Germans are probably more used to using English commands like
“/area list” instead of something like “/gebiet auflisten”, but since the
English commands should still work, even when language is set to German, I think
it’s okay to pull it through.

I further took the liberty to use “!status” as a command name instead of the
translation of “!tracked”, since it’s really about more than just tracked stuff
which relates to the “!track” command which in turn is all about Pokémon only.