improve everything-flag permission handling
So far, there were two option which controlled the “everything” flag:
- disableEverythingTracking, which (when true) disabled it altogether
- forceEverythingSeparately, which (when true) forced it to be resolved to
  single records in the database (that is: one per Pokémon respectively type of
  quest, instead of a wildcard-record), effectively enforcing an implicit
  “individually” flag

In addition it makes sense to allow people to deny the “individually” flag.

This commit implements that, and at the same time unifies all configuration
options into one with the following values and semantics:
- allow-any
  Allow either of the “everything” or “individually” flags.
- allow-and-always-individually
  Allow the “everything” flag and forcibly enable the “individually” flag.
- allow-and-ignore-individually
  Allow the “everything” flag and silently ignore the “individually” flag.
- deny
  Do not allow either of the “everything” or “individually” flags.

Any other value will result in “deny”.

For quests, “allow-individually-only” and “allow-everything-only” are identical
to “allow-any”, with the difference that there simply is no “individually” flag.