Fix min_weight and max_weight on monsters controller
data.weight was populated with values like 4.22, 5.60, 0.94 or 8.39
So when filtering for min_weight everything would go through the filter and when filtering for max_weight nothing would pass it. 
This works for my setup in stock RDM, it could be different in MAD. I would appreciate if someone could check it as I don't have access to MAD anymore.